Updating the report
Changed a script that was having issues when collecting perfmon data from FCI instances. To get the machine name and find the counter, instead of using SERVERPROPERTY('ComputerNamePhysicalNetBIOS'), it was using SERVERPROPERTY('MachineName'), which returns the SQL virtual name. The perfmon counters have the physical machine name always, even if this is a FCI.

Also changed a script that was causing issues when the instance where the nexus database is created doesn't have the collation  SQL_Latin1_General_CP1_CI_AS. Now it should work with any collation