ReadTrace post processing scripts added directly into sql nexus folder instead of generating dynamically (#473)
* #469 Inherit SQLScripts from Reporter and add hash for those

Removed the invoke of Reporter for scripts
Reformat scripts
Add files to the project
Add hash for these new scripts
First set of modifications,  this still needs additional testing.

* #469 Correcting some typos on the Script

#469 Correcting some typos on the Script

* #469 final modifications and test

+fixed minor typo issues
+removed the GO at the end , it was messing up batch parser and causing a fake failure
+Copilot confirmed old and new scripts are functionally same.
"only differences are in formatting, capitalization, and comments for readability. There are no changes to the logic, structure, or output of any procedures or functions."
+import tests